+++ app/src/main/java/com/devhub/moviesapp/domain/usecase/GetMovieCast.kt
@@ -9,7 +9,7 @@ import kotlinx.coroutines.flow.map
 class GetMovieCast(
     private val moviesRepository: MoviesRepository,
 ) {
-    operator fun invoke(movieId: Int): Flow<List<MovieCast>> {
+    operator fun invoke(movieId: Long): Flow<List<MovieCast>> {
         val listMovieCast = moviesRepository.getMovieCast(movieId).map { listMovieCastEntity ->
             listMovieCastEntity.map {
                 it.toMovieCast()
